fix room membership check on home page
Any authenticated user could previously access any room's data.
Fixed by:
- Adding validRoom() policy to page.jsx to redirect non-members
- Moving all Supabase queries out of the client into a new
  fetchRoomDashboard server action that validates membership before
  querying
- Converting HomeDashboard to a pure display component receiving
  data as props, removing direct DB access from the browser
